+++ src/chain/tezos/contracts/WrappedTezosHelper.ts
@@ -10,18 +10,16 @@ import { TezosConseilClient } from '../../../reporting/tezos/TezosConseilClient'
 import { ConseilServerInfo } from 'types/conseil/QueryTypes';
 import { ContractMapDetailsItem } from 'types/conseil/ConseilTezosTypes';
 import { TezosParameterFormat } from '../../../types/tezos/TezosChainTypes';
+import { StakerDaoTzip7 } from './StakerDaoTzip7';
 
 /** The expected checksum for the Wrapped Tezos contracts. */
 const CONTRACT_CHECKSUMS = {
-    token: 'd48b45bd77d2300026fe617c5ba7670e',
     oven: '5e3c30607da21a0fc30f7be61afb15c7',
     core: '7b9b5b7e7f0283ff6388eb783e23c452'
 }
 
 /** The expected checksum for the Wrapped Tezos scripts. */
 const SCRIPT_CHECKSUMS = {
-    // TODO(keefertaylor): Compute this checksum correctly.
-    token: '',
     // TODO(keefertaylor): Compute this checksum correctly.
     oven: '',
     // TODO(keefertaylor): Compute this checksum correctly.
@@ -38,22 +36,6 @@ export type OpenOvenResult = {
     // The address of the new oven contract.
     ovenAddress: string
 }
-
-export interface WrappedTezosStorage {
-    balanceMap: number;
-    approvalsMap: number;
-    supply: number;
-    administrator: string;
-    paused: boolean;
-    pauseGuardian: string;
-    outcomeMap: number;
-    swapMap: number;
-}
-
-export interface WrappedTezosBalanceRecord { }
-export interface WrappedTezosApprovalRecord { }
-export interface WrappedTezosOutcomeRecord { }
-export interface WrappedTezosSwapRecord { }
 
 /** 
  * Types for the Oven Map .
@@ -81,7 +63,7 @@ export type OvenMapSchema = { key: string, value: string }
  *
  * @author Keefer Taylor, Staker Services Ltd <keefer@stakerdao.com>
  */
-export namespace WrappedTezosHelper {
+namespace WrappedTezosHelperInternal {
     /**
      * Verifies that contract code for Wrapped Tezos matches the expected code.
      * 
@@ -99,7 +81,7 @@ export namespace WrappedTezosHelper {
         ovenContractAddress: string,
         coreContractAddress: string
     ): Promise<boolean> {
-        const tokenMatched = TezosContractUtils.verifyDestination(nodeUrl, tokenContractAddress, CONTRACT_CHECKSUMS.token)
+        const tokenMatched = StakerDaoTzip7.verifyDestination(nodeUrl, tokenContractAddress)
         const ovenMatched = TezosContractUtils.verifyDestination(nodeUrl, ovenContractAddress, CONTRACT_CHECKSUMS.oven)
         const coreMatched = TezosContractUtils.verifyDestination(nodeUrl, coreContractAddress, CONTRACT_CHECKSUMS.core)
 
@@ -121,97 +103,11 @@ export namespace WrappedTezosHelper {
         ovenScript: string,
         coreScript: string
     ): boolean {
-        const tokenMatched = TezosContractUtils.verifyScript(tokenScript, SCRIPT_CHECKSUMS.token)
+        const tokenMatched = StakerDaoTzip7.verifyScript(tokenScript)
         const ovenMatched = TezosContractUtils.verifyScript(ovenScript, SCRIPT_CHECKSUMS.oven)
         const coreMatched = TezosContractUtils.verifyScript(coreScript, SCRIPT_CHECKSUMS.core)
 
         return tokenMatched && ovenMatched && coreMatched
-    }
-
-    /**
-     *
-     * @param server
-     * @param address
-     */
-    export async function getSimpleStorage(server: string, address: string): Promise<WrappedTezosStorage> {
-        const storageResult = await TezosNodeReader.getContractStorage(server, address);
-
-        console.log(JSON.stringify(storageResult));
-
-        return {
-            balanceMap: Number(JSONPath({ path: '$.args[1].args[0].args[1].args[0].int', json: storageResult })[0]),
-            approvalsMap: Number(JSONPath({ path: '$.args[1].args[0].args[0].args[1].int', json: storageResult })[0]),
-            supply: Number(JSONPath({ path: '$.args[1].args[1].args[1].int', json: storageResult })[0]),
-            administrator: JSONPath({ path: '$.args[1].args[0].args[0].args[0].string', json: storageResult })[0],
-            paused: (JSONPath({ path: '$.args[1].args[1].args[0].prim', json: storageResult })[0]).toString().toLowerCase().startsWith('t'),
-            pauseGuardian: JSONPath({ path: '$.args[1].args[0].args[1].args[1].string', json: storageResult })[0],
-            outcomeMap: Number(JSONPath({ path: '$.args[0].args[0].int', json: storageResult })[0]),
-            swapMap: Number(JSONPath({ path: '$.args[0].args[1].int', json: storageResult })[0])
-        };
-    }
-
-    /**
-     * Get the balance of WXTZ tokens for an address.
-     * 
-     * @param nodeUrl The URL of the Tezos node which serves data.
-     * @param mapId The ID of the BigMap which contains balances.
-     * @param account The account to fetch the token balance for.
-     * @returns The balance of the account.
-     */
-    export async function getAccountBalance(server: string, mapid: number, account: string): Promise<number> {
-        const packedKey = TezosMessageUtils.encodeBigMapKey(Buffer.from(TezosMessageUtils.writePackedData(account, 'address'), 'hex'));
-        const mapResult = await TezosNodeReader.getValueForBigMapKey(server, mapid, packedKey);
-
-        if (mapResult === undefined) { throw new Error(`Map ${mapid} does not contain a record for ${account}`); }
-
-        const numberString = JSONPath({ path: '$.int', json: mapResult });
-        return Number(numberString);
-    }
-
-    /**
-     * Transfer some WXTZ between addresses.
-     * 
-     * @param nodeUrl The URL of the Tezos node which serves data.
-     * @param signer A Signer for the sourceAddress.
-     * @param keystore A Keystore for the sourceAddress.
-     * @param tokenContractAddress The address of the token contract. 
-     * @param fee The fee to use.
-     * @param sourceAddress The address which will send tokens.
-     * @param destinationAddress The address which will receive tokens.
-     * @param amount The amount of tokens to send.
-     * @param gasLimit The gas limit to use.
-     * @param storageLimit The storage limit to use. 
-     * @returns A string representing the operation hash.
-     */
-    export async function transferBalance(
-        nodeUrl: string,
-        signer: Signer,
-        keystore: KeyStore,
-        tokenContractAddress: string,
-        fee: number,
-        sourceAddress: string,
-        destinationAddress: string,
-        amount: number,
-        gasLimit: number = 51_300,
-        storageLimit: number = 70
-    ): Promise<string> {
-        const parameters = `Pair "${sourceAddress}" (Pair "${destinationAddress}" ${amount})`;
-
-        const nodeResult = await TezosNodeWriter.sendContractInvocationOperation(
-            nodeUrl,
-            signer,
-            keystore,
-            tokenContractAddress,
-            0,
-            fee,
-            storageLimit,
-            gasLimit,
-            'transfer',
-            parameters,
-            TezosTypes.TezosParameterFormat.Michelson
-        );
-
-        return TezosContractUtils.clearRPCOperationGroupHash(nodeResult.operationGroupID);
     }
 
     /**
@@ -497,3 +393,7 @@ export namespace WrappedTezosHelper {
         return TezosContractUtils.clearRPCOperationGroupHash(nodeResult.operationGroupID);
     }
 }
+
+/*** Combine namespaces */
+const WrappedTezosHelper = WrappedTezosHelperInternal || StakerDaoTzip7
+export WrappedTezosHelper
